added authentication to each end points

+++ TodoApp/routers/todos.py
@@ -2,14 +2,14 @@
 from typing import Annotated
 
 from pydantic import BaseModel, Field
-from starlette.status import HTTP_404_NOT_FOUND
+from starlette.status import HTTP_401_UNAUTHORIZED, HTTP_404_NOT_FOUND
 
 from sqlalchemy.orm import Session  # pyright: ignore[reportMissingImports]
 from fastapi import APIRouter, Depends, HTTPException, Query, status, Path
 
 from models import Todos
 from database import Base, SessionLocal, engine
-
+from .auth import get_current_user
 
 
 router = APIRouter()
@@ -24,6 +24,9 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session,  Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
+
+
 
 # Request model
 class TodoRequest(BaseModel):
@@ -32,33 +35,42 @@ class TodoRequest(BaseModel):
     priority: int = Field(gt=0, lt=6)
     complete: bool
 
+
 # Get all todos
 @router.get("/", status_code=status.HTTP_200_OK)
-def read_all(db: db_dependency):
-    return db.query(Todos).all()
+def read_all(db: db_dependency, user:user_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication  Failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 # Get todo by id
 @router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-def filter_with_id(todo_id: Annotated[int, Path(gt=0)], db: db_dependency):
-    todo = db.query(Todos).filter(Todos.id == todo_id).first()
+def filter_with_id(todo_id: Annotated[int, Path(gt=0)], user:user_dependency, db: db_dependency):
+    todo = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication  Failed")
     if todo is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     return todo
 
 
 # Create todo
-@router.post("/todos/createTodo", status_code=status.HTTP_201_CREATED)
-def create_todo(todo_add: TodoRequest, db:db_dependency):
-    todo_model = Todos(**todo_add.dict())
+@router.post("/createTodo", status_code=status.HTTP_201_CREATED)
+def create_todo(user: user_dependency, todo_add: TodoRequest, db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication  Failed")
+    todo_model = Todos(**todo_add.dict(), owner_id=user.get('id'))  # pyright: ignore[reportDeprecated]
     db.add(todo_model)
     db.commit()
 
 
 # Update todo
-@router.put("/todo/updateTodo/{todo_id}")
-def update_todo(todo_update: TodoRequest, db:db_dependency, todo_id: int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.put("/todo/updateTodo/{todo_id}", status_code=status.HTTP_202_ACCEPTED)
+def update_todo(todo_update: TodoRequest, db:db_dependency, user:user_dependency, todo_id: int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication  Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="todo not found")
     todo_model.title = todo_update.title
@@ -69,10 +81,12 @@ def update_todo(todo_update: TodoRequest, db:db_dependency, todo_id: int=Path(gt
     db.commit()
 
 # Delete todo
-@router.delete('/todo/{todo_id}')
-def delete_todo(db:db_dependency,todo_id: int=Path(gt=0)):
-    todo_model =db.query(Todos).filter(Todos.id ==todo_id).first()
+@router.delete('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
+def delete_todo(db:db_dependency, user:user_dependency, todo_id: int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication  Failed")
+    todo_model =db.query(Todos).filter(Todos.id ==todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('id')).delete()
     db.commit()
